Initial refactor: separate pool and provider
