+++ blockchain/message_pool/src/lib.rs
@@ -11,3 +11,5 @@ pub use self::block_prob::*;
 pub use self::config::*;
 pub use self::errors::*;
 pub use self::msgpool::*;
+pub use self::msgpool::msg_pool::MessagePool;
+pub use self::msgpool::provider::MpoolRpcProvider;
