+++ blockchain/message_pool/src/msgpool/mod.rs
@@ -2,7 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 
 mod selection;
-use super::config::MpoolConfig;
+pub(crate) mod msg_pool;
+pub(crate) mod provider;
+pub(crate) mod utils;
+
 use super::errors::Error;
 use crate::msg_chain::{MsgChain, MsgChainNode};
 use address::{Address, Protocol};
@@ -35,6 +38,7 @@ use std::time::Duration;
 use std::{borrow::BorrowMut, cmp::Ordering};
 use tokio::sync::broadcast::{error::RecvError, Receiver as Subscriber, Sender as Publisher};
 use types::verifier::ProofVerifier;
+use utils::{get_base_fee_lower_bound, get_gas_reward, get_gas_perf};
 use vm::ActorState;
 
 const REPLACE_BY_FEE_RATIO: f32 = 1.25;
@@ -161,580 +165,6 @@ pub trait Provider {
     fn chain_compute_base_fee(&self, ts: &Tipset) -> Result<BigInt, Error>;
 }
 
-/// This is the default Provider implementation that will be used for the mpool RPC.
-pub struct MpoolRpcProvider<DB> {
-    subscriber: Publisher<HeadChange>,
-    sm: Arc<StateManager<DB>>,
-}
-
-impl<DB> MpoolRpcProvider<DB>
-where
-    DB: BlockStore + Sync + Send,
-{
-    pub fn new(subscriber: Publisher<HeadChange>, sm: Arc<StateManager<DB>>) -> Self
-    where
-        DB: BlockStore,
-    {
-        MpoolRpcProvider { subscriber, sm }
-    }
-}
-
-#[async_trait]
-impl<DB> Provider for MpoolRpcProvider<DB>
-where
-    DB: BlockStore + Sync + Send + 'static,
-{
-    async fn subscribe_head_changes(&mut self) -> Subscriber<HeadChange> {
-        self.subscriber.subscribe()
-    }
-
-    async fn get_heaviest_tipset(&mut self) -> Option<Arc<Tipset>> {
-        self.sm.chain_store().heaviest_tipset().await
-    }
-
-    fn put_message(&self, msg: &ChainMessage) -> Result<Cid, Error> {
-        let cid = self
-            .sm
-            .blockstore()
-            .put(msg, Blake2b256)
-            .map_err(|err| Error::Other(err.to_string()))?;
-        Ok(cid)
-    }
-
-    fn get_actor_after(&self, addr: &Address, ts: &Tipset) -> Result<ActorState, Error> {
-        let state = StateTree::new_from_root(self.sm.blockstore(), ts.parent_state())
-            .map_err(|e| Error::Other(e.to_string()))?;
-        let actor = state
-            .get_actor(addr)
-            .map_err(|e| Error::Other(e.to_string()))?;
-        actor.ok_or_else(|| Error::Other("No actor state".to_owned()))
-    }
-
-    fn messages_for_block(
-        &self,
-        h: &BlockHeader,
-    ) -> Result<(Vec<UnsignedMessage>, Vec<SignedMessage>), Error> {
-        chain::block_messages(self.sm.blockstore(), h).map_err(|err| err.into())
-    }
-
-    fn messages_for_tipset(&self, h: &Tipset) -> Result<Vec<ChainMessage>, Error> {
-        Ok(self.sm.chain_store().messages_for_tipset(h)?)
-    }
-
-    async fn load_tipset(&self, tsk: &TipsetKeys) -> Result<Arc<Tipset>, Error> {
-        let ts = self.sm.chain_store().tipset_from_keys(tsk).await?;
-        Ok(ts)
-    }
-    fn chain_compute_base_fee(&self, ts: &Tipset) -> Result<BigInt, Error> {
-        chain::compute_base_fee(self.sm.blockstore(), ts).map_err(|err| err.into())
-    }
-    async fn state_account_key<V>(&self, addr: &Address, ts: &Arc<Tipset>) -> Result<Address, Error>
-    where
-        V: ProofVerifier,
-    {
-        self.sm
-            .resolve_to_key_addr::<V>(addr, ts)
-            .await
-            .map_err(|e| Error::Other(e.to_string()))
-    }
-}
-
-/// This contains all necessary information needed for the message pool.
-/// Keeps track of messages to apply, as well as context needed for verifying transactions.
-pub struct MessagePool<T> {
-    local_addrs: Arc<RwLock<Vec<Address>>>,
-    pending: Arc<RwLock<HashMap<Address, MsgSet>>>,
-    pub cur_tipset: Arc<RwLock<Arc<Tipset>>>,
-    api: Arc<RwLock<T>>,
-    pub min_gas_price: BigInt,
-    pub max_tx_pool_size: i64,
-    pub network_name: String,
-    network_sender: Sender<NetworkMessage>,
-    bls_sig_cache: Arc<RwLock<LruCache<Cid, Signature>>>,
-    sig_val_cache: Arc<RwLock<LruCache<Cid, ()>>>,
-    republished: Arc<RwLock<HashSet<Cid>>>,
-    repub_trigger: Sender<()>,
-    // TODO look into adding a cap to local_msgs
-    local_msgs: Arc<RwLock<HashSet<SignedMessage>>>,
-    config: MpoolConfig,
-}
-
-impl<T> MessagePool<T>
-where
-    T: Provider + std::marker::Send + std::marker::Sync + 'static,
-{
-    pub async fn new(
-        mut api: T,
-        network_name: String,
-        network_sender: Sender<NetworkMessage>,
-        config: MpoolConfig,
-    ) -> Result<MessagePool<T>, Error>
-    where
-        T: Provider,
-    {
-        let local_addrs = Arc::new(RwLock::new(Vec::new()));
-        let pending = Arc::new(RwLock::new(HashMap::new()));
-        let tipset = Arc::new(RwLock::new(api.get_heaviest_tipset().await.ok_or_else(
-            || Error::Other("Failed to retrieve heaviest tipset from provider".to_owned()),
-        )?));
-        // LruCache sizes have been taken from the lotus implementation
-        let bls_sig_cache = Arc::new(RwLock::new(LruCache::new(40000)));
-        let sig_val_cache = Arc::new(RwLock::new(LruCache::new(32000)));
-        let api_mutex = Arc::new(RwLock::new(api));
-        let local_msgs = Arc::new(RwLock::new(HashSet::new()));
-        let republished = Arc::new(RwLock::new(HashSet::new()));
-
-        let (repub_trigger, mut repub_trigger_rx) = bounded::<()>(4);
-        let mut mp = MessagePool {
-            local_addrs,
-            pending,
-            cur_tipset: tipset,
-            api: api_mutex,
-            min_gas_price: Default::default(),
-            max_tx_pool_size: 5000,
-            network_name,
-            bls_sig_cache,
-            sig_val_cache,
-            local_msgs,
-            republished,
-            config,
-            network_sender,
-            repub_trigger,
-        };
-
-        mp.load_local().await?;
-
-        let mut subscriber = mp.api.write().await.subscribe_head_changes().await;
-
-        let api = mp.api.clone();
-        let bls_sig_cache = mp.bls_sig_cache.clone();
-        let pending = mp.pending.clone();
-        let republished = mp.republished.clone();
-
-        let cur_tipset = mp.cur_tipset.clone();
-        let repub_trigger = Arc::new(mp.repub_trigger.clone());
-
-        // Reacts to new HeadChanges
-        task::spawn(async move {
-            loop {
-                match subscriber.recv().await {
-                    Ok(ts) => {
-                        let (cur, rev, app) = match ts {
-                            HeadChange::Current(_tipset) => continue,
-                            HeadChange::Revert(tipset) => (
-                                cur_tipset.clone(),
-                                vec![tipset.as_ref().clone()],
-                                Vec::new(),
-                            ),
-                            HeadChange::Apply(tipset) => (
-                                cur_tipset.clone(),
-                                Vec::new(),
-                                vec![tipset.as_ref().clone()],
-                            ),
-                        };
-                        head_change(
-                            api.as_ref(),
-                            bls_sig_cache.as_ref(),
-                            repub_trigger.clone(),
-                            republished.as_ref(),
-                            pending.as_ref(),
-                            &cur.as_ref(),
-                            rev,
-                            app,
-                        )
-                        .await
-                        .unwrap_or_else(|err| warn!("Error changing head: {:?}", err));
-                    }
-                    Err(RecvError::Lagged(e)) => {
-                        log::warn!("Head change subscriber lagged: skipping {} events", e);
-                    }
-                    Err(RecvError::Closed) => {
-                        break;
-                    }
-                }
-            }
-        });
-
-        let api = mp.api.clone();
-        let pending = mp.pending.clone();
-        let cur_tipset = mp.cur_tipset.clone();
-        let republished = mp.republished.clone();
-        let local_addrs = mp.local_addrs.clone();
-        let network_sender = Arc::new(mp.network_sender.clone());
-        let network_name = mp.network_name.clone();
-        // Reacts to republishing requests
-        task::spawn(async move {
-            let mut interval = interval(Duration::from_millis(REPUBLISH_INTERVAL));
-            loop {
-                select(interval.next(), repub_trigger_rx.next()).await;
-                if let Err(e) = republish_pending_messages(
-                    api.as_ref(),
-                    network_sender.as_ref(),
-                    network_name.as_ref(),
-                    pending.as_ref(),
-                    cur_tipset.as_ref(),
-                    republished.as_ref(),
-                    local_addrs.as_ref(),
-                )
-                .await
-                {
-                    warn!("Failed to republish pending messages: {}", e.to_string());
-                }
-            }
-        });
-        Ok(mp)
-    }
-
-    /// Add a signed message to local_addrs and local_msgs.
-    async fn add_local(&self, m: SignedMessage) -> Result<(), Error> {
-        self.local_addrs.write().await.push(*m.from());
-        self.local_msgs.write().await.insert(m);
-        Ok(())
-    }
-
-    /// Push a signed message to the MessagePool.
-    pub async fn push(&self, msg: SignedMessage) -> Result<Cid, Error> {
-        self.check_message(&msg).await?;
-        let cid = msg.cid().map_err(|err| Error::Other(err.to_string()))?;
-        let cur_ts = self.cur_tipset.read().await.clone();
-        let publish = self.add_tipset(msg.clone(), &cur_ts, true).await?;
-        let msg_ser = msg.marshal_cbor()?;
-        self.add_local(msg).await?;
-        if publish {
-            self.network_sender
-                .send(NetworkMessage::PubsubMessage {
-                    topic: Topic::new(format!("{}/{}", PUBSUB_MSG_STR, self.network_name)),
-                    message: msg_ser,
-                })
-                .await
-                .map_err(|_| Error::Other("Network receiver dropped".to_string()))?;
-        }
-        Ok(cid)
-    }
-
-    /// Basic checks on the validity of a message.
-    async fn check_message(&self, msg: &SignedMessage) -> Result<(), Error> {
-        if msg.marshal_cbor()?.len() > 32 * 1024 {
-            return Err(Error::MessageTooBig);
-        }
-        msg.message()
-            .valid_for_block_inclusion(0, NEWEST_NETWORK_VERSION)
-            .map_err(Error::Other)?;
-        if msg.value() > &types::TOTAL_FILECOIN {
-            return Err(Error::MessageValueTooHigh);
-        }
-        if msg.gas_fee_cap() < &MINIMUM_BASE_FEE {
-            return Err(Error::GasFeeCapTooLow);
-        }
-        self.verify_msg_sig(msg).await
-    }
-
-    /// This is a helper to push that will help to make sure that the message fits the parameters
-    /// to be pushed to the MessagePool.
-    pub async fn add(&self, msg: SignedMessage) -> Result<(), Error> {
-        self.check_message(&msg).await?;
-
-        let tip = self.cur_tipset.read().await.clone();
-
-        self.add_tipset(msg, &tip, false).await?;
-        Ok(())
-    }
-
-    /// Add a SignedMessage without doing any of the checks.
-    pub async fn add_skip_checks(&mut self, m: SignedMessage) -> Result<(), Error> {
-        self.add_helper(m).await
-    }
-
-    /// Verify the message signature. first check if it has already been verified and put into
-    /// cache. If it has not, then manually verify it then put it into cache for future use.
-    async fn verify_msg_sig(&self, msg: &SignedMessage) -> Result<(), Error> {
-        let cid = msg.cid()?;
-
-        if let Some(()) = self.sig_val_cache.write().await.get(&cid) {
-            return Ok(());
-        }
-
-        msg.verify().map_err(Error::Other)?;
-
-        self.sig_val_cache.write().await.put(cid, ());
-
-        Ok(())
-    }
-
-    /// Verify the state_sequence and balance for the sender of the message given then
-    /// call add_locked to finish adding the signed_message to pending.
-    async fn add_tipset(
-        &self,
-        msg: SignedMessage,
-        cur_ts: &Tipset,
-        local: bool,
-    ) -> Result<bool, Error> {
-        let sequence = self.get_state_sequence(msg.from(), cur_ts).await?;
-
-        if sequence > msg.message().sequence() {
-            return Err(Error::SequenceTooLow);
-        }
-
-        let publish = verify_msg_before_add(&msg, &cur_ts, local)?;
-
-        let balance = self.get_state_balance(msg.from(), cur_ts).await?;
-
-        let msg_balance = msg.message().required_funds();
-        if balance < msg_balance {
-            return Err(Error::NotEnoughFunds);
-        }
-        self.add_helper(msg).await?;
-        Ok(publish)
-    }
-
-    /// Finish verifying signed message before adding it to the pending mset hashmap. If an entry
-    /// in the hashmap does not yet exist, create a new mset that will correspond to the from
-    /// message and push it to the pending hashmap.
-    async fn add_helper(&self, msg: SignedMessage) -> Result<(), Error> {
-        let from = *msg.from();
-        let cur_ts = self.cur_tipset.read().await.clone();
-        add_helper(
-            self.api.as_ref(),
-            self.bls_sig_cache.as_ref(),
-            self.pending.as_ref(),
-            msg,
-            self.get_state_sequence(&from, &cur_ts).await?,
-        )
-        .await
-    }
-
-    /// Get the sequence for a given address, return Error if there is a failure to retrieve
-    /// the respective sequence.
-    pub async fn get_sequence(&self, addr: &Address) -> Result<u64, Error> {
-        let cur_ts = self.cur_tipset.read().await.clone();
-
-        let sequence = self.get_state_sequence(addr, &cur_ts).await?;
-
-        let pending = self.pending.read().await;
-
-        let msgset = pending.get(addr);
-        match msgset {
-            Some(mset) => {
-                if sequence > mset.next_sequence {
-                    return Ok(sequence);
-                }
-                Ok(mset.next_sequence)
-            }
-            None => Ok(sequence),
-        }
-    }
-
-    /// Get the state of the sequence for a given address in cur_ts.
-    async fn get_state_sequence(&self, addr: &Address, cur_ts: &Tipset) -> Result<u64, Error> {
-        let actor = self.api.read().await.get_actor_after(&addr, cur_ts)?;
-        Ok(actor.sequence)
-    }
-
-    /// Get the state balance for the actor that corresponds to the supplied address and tipset,
-    /// if this actor does not exist, return an error.
-    async fn get_state_balance(&self, addr: &Address, ts: &Tipset) -> Result<BigInt, Error> {
-        let actor = self.api.read().await.get_actor_after(&addr, &ts)?;
-        Ok(actor.balance)
-    }
-
-    /// Adds a local message returned from the call back function with the current nonce.
-    pub async fn push_with_sequence<V>(&self, addr: &Address, cb: T) -> Result<SignedMessage, Error>
-    where
-        T: Fn(Address, u64) -> Result<SignedMessage, Error>,
-        V: ProofVerifier,
-    {
-        let cur_ts = self.cur_tipset.read().await.clone();
-        let from_key = match addr.protocol() {
-            Protocol::ID => {
-                let api = self.api.read().await;
-
-                api.state_account_key::<V>(&addr, &self.cur_tipset.read().await.clone())
-                    .await?
-            }
-            _ => *addr,
-        };
-
-        let sequence = self.get_sequence(&addr).await?;
-        let msg = cb(from_key, sequence)?;
-        self.check_message(&msg).await?;
-        if *self.cur_tipset.read().await != cur_ts {
-            return Err(Error::TryAgain);
-        }
-
-        if self.get_sequence(&addr).await? != sequence {
-            return Err(Error::TryAgain);
-        }
-
-        let publish = verify_msg_before_add(&msg, &cur_ts, true)?;
-        self.check_balance(&msg, &cur_ts).await?;
-        self.add_helper(msg.clone()).await?;
-        self.add_local(msg.clone()).await?;
-
-        if publish {
-            self.network_sender
-                .send(NetworkMessage::PubsubMessage {
-                    topic: Topic::new(format!("{}/{}", PUBSUB_MSG_STR, self.network_name)),
-                    message: msg.marshal_cbor()?,
-                })
-                .await
-                .map_err(|_| Error::Other("Network receiver dropped".to_string()))?;
-        }
-
-        Ok(msg)
-    }
-
-    async fn check_balance(&self, m: &SignedMessage, cur_ts: &Tipset) -> Result<(), Error> {
-        let bal = self.get_state_balance(m.from(), &cur_ts).await?;
-        let mut required_funds = m.required_funds();
-        if bal < required_funds {
-            return Err(Error::NotEnoughFunds);
-        }
-        if let Some(mset) = self.pending.read().await.get(m.from()) {
-            required_funds += mset.get_required_funds(m.sequence());
-        }
-        if bal < required_funds {
-            return Err(Error::SoftValidationFailure(format!(
-                "not enough funds including pending messages (required: {}, balance: {})",
-                required_funds, bal
-            )));
-        }
-        Ok(())
-    }
-
-    /// Remove a message given a sequence and address from the messagepool.
-    pub async fn remove(
-        &mut self,
-        from: &Address,
-        sequence: u64,
-        applied: bool,
-    ) -> Result<(), Error> {
-        remove(from, self.pending.as_ref(), sequence, applied).await
-    }
-
-    /// Return a tuple that contains a vector of all signed messages and the current tipset for
-    /// self.
-    pub async fn pending(&self) -> Result<(Vec<SignedMessage>, Arc<Tipset>), Error> {
-        let mut out: Vec<SignedMessage> = Vec::new();
-        let pending = self.pending.read().await;
-        let pending_hm = pending.clone();
-
-        for (addr, _) in pending_hm {
-            out.append(
-                self.pending_for(&addr)
-                    .await
-                    .ok_or(Error::InvalidFromAddr)?
-                    .as_mut(),
-            )
-        }
-
-        let cur_ts = self.cur_tipset.read().await.clone();
-
-        Ok((out, cur_ts))
-    }
-
-    /// Return a Vector of signed messages for a given from address. This vector will be sorted by
-    /// each messsage's sequence. If no corresponding messages found, return None result type.
-    pub async fn pending_for(&self, a: &Address) -> Option<Vec<SignedMessage>> {
-        let pending = self.pending.read().await;
-        let mset = pending.get(a)?;
-        if mset.msgs.is_empty() {
-            return None;
-        }
-        let mut msg_vec = Vec::new();
-        for (_, item) in mset.msgs.iter() {
-            msg_vec.push(item.clone());
-        }
-        msg_vec.sort_by_key(|value| value.message().sequence());
-        Some(msg_vec)
-    }
-
-    /// Return Vector of signed messages given a block header for self.
-    pub async fn messages_for_blocks(
-        &self,
-        blks: &[BlockHeader],
-    ) -> Result<Vec<SignedMessage>, Error> {
-        let mut msg_vec: Vec<SignedMessage> = Vec::new();
-
-        for block in blks {
-            let (umsg, mut smsgs) = self.api.read().await.messages_for_block(&block)?;
-
-            msg_vec.append(smsgs.as_mut());
-            for msg in umsg {
-                let mut bls_sig_cache = self.bls_sig_cache.write().await;
-                let smsg = recover_sig(&mut bls_sig_cache, msg).await?;
-                msg_vec.push(smsg)
-            }
-        }
-        Ok(msg_vec)
-    }
-
-    /// Return gas price estimate this has been translated from lotus, a more smart implementation will
-    /// most likely need to be implemented.
-    // TODO: UPDATE https://github.com/ChainSafe/forest/issues/901
-    pub fn estimate_gas_premium(
-        &self,
-        nblocksincl: u64,
-        _sender: Address,
-        _gas_limit: u64,
-        _tsk: TipsetKeys,
-    ) -> Result<BigInt, Error> {
-        let min_gas_price = 0;
-        match nblocksincl {
-            0 => Ok(BigInt::from(min_gas_price + 2)),
-            1 => Ok(BigInt::from(min_gas_price + 1)),
-            _ => Ok(BigInt::from(min_gas_price)),
-        }
-    }
-
-    /// Loads local messages to the message pool to be applied.
-    pub async fn load_local(&mut self) -> Result<(), Error> {
-        let mut local_msgs = self.local_msgs.write().await;
-        let msg_vec: Vec<SignedMessage> = local_msgs.iter().cloned().collect();
-
-        for k in msg_vec.into_iter() {
-            self.add(k.clone()).await.unwrap_or_else(|err| {
-                if err == Error::SequenceTooLow {
-                    warn!("error adding message: {:?}", err);
-                    local_msgs.remove(&k);
-                }
-            })
-        }
-
-        Ok(())
-    }
-    /// If `local = true`, the local messages will be removed as well as pending messages.
-    /// If `local = false`, pending messages will be removed while retaining local messages.
-    pub async fn clear(&mut self, local: bool) {
-        if local {
-            let local_addrs = self.local_addrs.read().await;
-            for a in local_addrs.iter() {
-                if let Some(mset) = self.pending.read().await.get(&a) {
-                    for m in mset.msgs.values() {
-                        if !self.local_msgs.write().await.remove(&m) {
-                            warn!("error deleting local message");
-                        }
-                    }
-                }
-            }
-            self.pending.write().await.clear();
-            self.republished.write().await.clear();
-        } else {
-            let mut pending = self.pending.write().await;
-            let local_addrs = self.local_addrs.read().await;
-            pending.retain(|a, _| local_addrs.contains(&a));
-        }
-    }
-    pub fn get_config(&self) -> &MpoolConfig {
-        &self.config
-    }
-    pub fn set_config<DB: Store>(&mut self, db: &DB, cfg: MpoolConfig) -> Result<(), Error> {
-        cfg.save_config(db)
-            .map_err(|e| Error::Other(e.to_string()))?;
-        self.config = cfg;
-        Ok(())
-    }
-}
-
 fn verify_msg_before_add(m: &SignedMessage, cur_ts: &Tipset, local: bool) -> Result<bool, Error> {
     let epoch = cur_ts.epoch();
     let min_gas = interpreter::price_list_by_epoch(epoch).on_chain_message(m.marshal_cbor()?.len());
@@ -756,14 +186,6 @@ fn verify_msg_before_add(m: &SignedMessage, cur_ts: &Tipset, local: bool) -> Res
         }
     }
     Ok(local)
-}
-
-fn get_base_fee_lower_bound(base_fee: &BigInt, factor: i64) -> BigInt {
-    let base_fee_lower_bound = base_fee.div_floor(&BigInt::from(factor));
-    if base_fee_lower_bound < *MINIMUM_BASE_FEE {
-        return MINIMUM_BASE_FEE.clone();
-    }
-    base_fee_lower_bound
 }
 
 /// Remove a message from pending given the from address and sequence.
@@ -848,6 +270,7 @@ where
 
     Ok(())
 }
+
 /// Get the state of the base_sequence for a given address in cur_ts.
 async fn get_state_sequence<T>(
     api: &RwLock<T>,
@@ -974,6 +397,7 @@ where
 
     Ok(())
 }
+
 async fn create_message_chains<T>(
     api: &RwLock<T>,
     actor: &Address,
@@ -1126,20 +550,6 @@ where
     }
     // No need to link the chains because its linked for free
     Ok(chains)
-}
-
-/// Gets the gas reward for the given message.
-pub(crate) fn get_gas_reward(msg: &SignedMessage, base_fee: &BigInt) -> BigInt {
-    let mut max_prem = msg.gas_fee_cap() - base_fee;
-    if &max_prem < msg.gas_premium() {
-        max_prem = msg.gas_premium().clone();
-    }
-    max_prem * msg.gas_limit()
-}
-
-pub(crate) fn get_gas_perf(gas_reward: &BigInt, gas_limit: i64) -> f64 {
-    let a = BigRational::new(gas_reward * types::BLOCK_GAS_LIMIT, gas_limit.into());
-    a.to_f64().unwrap()
 }
 
 /// This function will revert and/or apply tipsets to the message pool. This function should be
@@ -1561,7 +971,7 @@ pub mod test_provider {
 pub mod tests {
     use super::test_provider::*;
     use super::*;
-    use crate::MessagePool;
+    use crate::msg_pool::MessagePool;
     use address::Address;
     use async_std::channel::bounded;
     use async_std::task;
